fix: Addressing review comments - extracted common tracer logic in parent class
